+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Finally add a CHANGELOG.md to communicate breaking changes (#598)
 - Test multiple versions of InfluxDB in travis
+- Add SHARD DURATION parameter to retention policy create/alter
 ### Changed
 - Update POST/GET requests to follow verb guidelines from InfluxDB documentation
 - Update test suite to support InfluxDB v1.3.9, v1.4.2, and v1.5.4